moved rebalance validators
